fix: refactored into multiple controllers
